Move a few functions that have been copy-pasted around to shared module.
